Updated default.toml files are seen on package update
Previously, if a `default.toml` file changed content or structure
between releases of a package, this difference was not picked up until
the Supervisor was restarted (specifically, until the internal
`Service` record was dropped and recreated).

Now, when a package is updated, we compare the `default.toml` contents
of the incoming release with our current default configuration
layer. If they differ, we set a flag on the `Service` struct; this is
now incorporated into our logic that determines if the configuration
as a whole has changed.

As a result, templates will now be re-rendered according to updated
`default.toml` content.

Fixes #4576

Signed-off-by: Christopher Maier <cmaier@chef.io>