Formalize template rendering context
Previously, the data we made available to users in Habitat templates
was based directly on the JSON serialization of several internal data
structures. This needlessly coupled internal workings of Habitat with
the public interface we expose to users. This often made it difficult
to know exactly what data was presented to users, and made it
difficult to safely change that data.

Here, we explicitly decouple the internal data structures of the
Supervisor from what is exposed to users. To do this in an efficient
and clear way, we introduce a number of "templating proxies", which
effectively wrap internal data structures, holding on to their
relevant data using `Cow` types to reduce memory footprint (this also
makes it easier to instantiate the templating proxies in tests, since
we can simply provide owned copies of the important data). Not only
does this give us a clear, typed way to refer to the data, but it
allows us to implement `Serialize` on these types, which lets us
control what is exposed to users in a very fine-grained and targeted
way. For instance, `{{svc.me.persistent}}` should really be
`{{svc.me.permanent}}`; before this change, we would need to introduce
another field to `census::CensusMember` named `permanent`, and it
would be unclear why. Now, however, we can simply modify the
templating proxy's `Serialize` implementation to render that same
field multiple times under different names, making schema evolution
and deprecation much easier to reason about. Note, however, that no
user-facing changes are introduced in this commit; it is a pure
refactoring.

By clearly separating out the templating data, we now have an
opportunity to revisit some of the internal data structures that were
previously serving "double duty"; it may be possible to simplify them
now.

Additionally, this commit introduces a JSON Schema definition of the
entire render context. This is currently used in tests to validate our
code, but could be used in the future as a source for generated
documentation on the website, and possibly as part of a future
template validation tool.

Fixes #4761

Signed-off-by: Christopher Maier <cmaier@chef.io>