Remove Heap script
It appears all we need here is the Segment script, since the write key we’re using with it is wired up to funnel usage details into Heap on the receiving end. This change removes Heap in order to match up with what we’re doing as of yesterday in Builder.

Signed-off-by: Christian Nunciato <chris@nunciato.org>