Revert to boolean-based health in CensusMember
This fixes a subtlety introduced with the addition of a `Health` field
to the `CensusMember` struct (PR #4823). Previously, there was a small
window where the health of a `CensusMember` was in an indeterminate
state, before it's health was reconciled with Butterfly data. By using
the `Health` enum, all new `CensusMember`s were created as `Alive`.

Arguments could be made for the correctness or incorrectness of this
decision, and we're all agreed that having four booleans to represent
this data is not what we ultimately want. However, since we don't want
to introduce subtle bugs into this part of the code in the course of
solidifying our template data structures, we're backing this change
out for now.

Signed-off-by: Christopher Maier <cmaier@chef.io>