Expose health-check status in HTTP gateway response
We currently expose service health in the HTTP gateway at
`http://<SERVER>:9631/services/<SERVICE>/<GROUP>/health`. The HTTP
status of the response is derived from the health check status,
according to the following mapping:

OK, WARNING => HTTP 200
CRITICAL => HTTP 503
UNKNOWN => HTTP 500

The contents of the `stdout` and `stderr` streams of the health check
script are returned in the body of the response.

This works well, as long as you don't need to distinguish between the
OK and WARNING states.

Since there is no immediately obvious HTTP status that might
correspond with a WARNING state, this commit adds a `status` field to
the JSON response body, the value of which is either "OK", "WARNING",
"CRITICAL", or "UNKNOWN". Users that need to distinguish between OK
and WARNING can inspect the body of HTTP 200 responses to determine
which state the service is currently in.

Signed-off-by: Christopher Maier <cmaier@chef.io>