Validate configuration before applying to service group
Apply a basic validation against a running service's configuration
interface. Using `default.toml` (for now) we can message back to the
user that an unknown key was set. This validation can, and will be
expanded upon as we expand upon our configuration interface by
converting `default.toml` from a simple key value store to a markup
which will enable better validations. For example:

```
[my_value]
this = 1
```

Could become:

```
[my_value]
type = "integer"
required = true
default = 1
```

This is now possible because of the addition of the CtlGateway. With
it, we are able to communicate in a request response pattern with a
Supervisor that is aware of what it believes the state of the ring
is and it's own state.

* Send configuration through the Control Gateway and then into the
  gossip ring once validated
* Move `config apply` from hab-butterfly to hab-sup

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>