[core] Restore impl for PackageInstall#legacy_runtime_path().
This change restores a previous implementation of the
`PackageInstall#runtime_path()` method but renames it to
`PackageInstall#legacy_runtime_path()` to be used by older packages
which lack the `RUNTIME_ENVIRONMENT` metafile.

When verifying the runtime path ordering in the previous commit, the
calculated path ordering wasn't correct with respect to direct
dependencies vs. transitive dependencies.

The preserved reference implementation can be found in an older version
of the `habitat-sh/habitat` repository:

https://github.com/habitat-sh/habitat/blob/333b75d6234db0531cf4a5bdcb859f7d4adc2478/components/core/src/package/install.rs#L321-L350

In short, the path ordering needs to be:

1. All path entries from the package itself
2. All path entries (ordered as read from the `PATH` metafile) from each
direct dependency, ordered as read from the `DEPS` metafile
3. All path entries (ordered as read from the `PATH` metafile) from any
remaining transitive dependency, ordered as read from the `TDEPS`
metafile (which is lexically sorted at build time)

Ideally there shouldn't be any path entry duplications, but the ordering
is more important than any duplication. The corect ordering ensures that
the earlier entries in a Plan's `$pkg_deps` will be on `$PATH` first
(and therefore will "win" if another program of the same name comes from
another package also on `$PATH`).

A net-new unit test was added in this change to ensure the correct
behavior in future feature work.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>