Present only active group members in the RenderContext
Previously, we provided *all* members of a group to the
`RenderContext`. Since we don't yet do any rumor cleanup, this
included all the confirmed-dead and departed members of the group;
depending on how long the group has been in existence, this could make
up the overwhelming majority of the list. I'm hard-pressed to come up
with a legitimate scenario where you would want access to dead members
in a template. This means that users pretty much had to always use the
`eachAlive` helper function to filter the list, but that still makes
it hard to have logic that hinges on whether or not there are any
active members around. Furthermore, it takes time and memory to
process and render all those extra dead census members into JSON.

Now, all `svc.members` and `bind.<SERVICE>.members` templating calls
return only members that are either alive or suspect. This means that
users will no longer need to use the `eachAlive` helper function, and
can even start using Handlebars testing contstructs like

    {{ #if bind.<SERVICE>.members }}

We return suspect members as well as live ones because the suspicion
mechanism is a feature of the underlying membership protocol that
counteracts false-negatives. A member may be suspect due to momentary
network congestion, and altering your templating based on that could
result in unnecessary flapping of your service.

Note that this is _slightly_ different than if you were using the
`eachAlive` helper (since that does not include suspect members), but
it is arguably more correct.

Since we're potentially filtering out a large number of members this
way, this change will likely reduce memory consumption, as well as
speed up rendering, even if only slightly.

Signed-off-by: Christopher Maier <cmaier@chef.io>