Update schema to be in line with underlying Rust implementation.
Based on the current implementation, it _shouldn't_ be possible to get
a bind that has no `first` member, even though it is technically
possible. Since `first` is deprecated now anyway, we'll just maintain
the code as-is, bring the JSON schema in line with it, and just yank
it all out later.

Signed-off-by: Christopher Maier <cmaier@chef.io>