[hab-sup] Exit 0 when given --help or --version
Previously, hab-sup --version would exit with 86. Now, we exit 0 for
both --version and --help.

Note, however, we will still exit with an error if the help is
displayed because of an error such as a bad subcommand or bad option.

You may ask, do we need to print the error in the case of
HelpDisplayed or VersionDisplayed. In the case of HelpDisplayed the
answer is yes because the help output isn't actually output to stdout
but rather to a buffer that is contained in the error. In the case of
VersionDisplay the answer is "kinda". The version is printed to
stdout but without a newline, so printing the empty error prints a
newline, which is a bit dirty but kept the function simple.

Signed-off-by: Steven Danna <steve@chef.io>