[plan-build] Treat PATH distinctly from RUNTIME_ENVIRONMENT.
This change further re-asserts how the PATH environment variable is used
and interpreted in difference places. In the context of the build
program, there are 3 primary places where we care about something called
`$PATH`:

1. The path elements in a package which contain programs, declared by a
Plan author in `$pkg_bin_dirs[@]` and written out to a `PATH` metadata
file.
2. The "runtime" value for `$PATH` which the Supervisor (and others) use
when starting a service. This is a combination of a package's own path
elements, and those of its direct and indirect runtime dependencies
(ordered in a particular way). It is written out to a `PATH` key in a
`RUNTIME_ENVIRONMENT` metadata file.
3. The value of `$PATH` which is exported for the build program to use
when building the package. This is only used in the running context of
the build program and contains build dependencies as well as runtime
dependencies.

This change deals with cases 2 and 3 and leaves 1 as-is.

2. Runtime Environment PATH
---------------------------

Prior to this change, the value of the `PATH` key in the
`RUNTIME_ENVIRONMENT` metadata file was calculated by pulling in the
`RUNTIME_ENVIRONMENT` from dependencies and layering them on themselves.
The issue was it was to use this information to derive a clearly ordered
path array that is in the following order:

1. All directories in this package which have program binaries (set by
the Plan author in `${pkg_bin_dirs[@]}`), in declared order
2. All path elements for each *direct* runtime dependency which has a
`PATH` metadata file, in the order present in `PATH`
3. All remaining path elements for each *indirect* (i.e. transitive)
dependency which has a `PATH` metadata file, in the order present in
`PATH`

Ideally, this collection should be a unique ordered set with no
duplicated path elements.

It ended up being easier to re-compute this runtime PATH ordering by
querying the direct and indirect dependencies directly and derive the
result without using the `RUNTIME_ENVIRONMENT` metadata files. This
seemed to suggest that the notion of "PATH" is more like the `CFLAGS`
and `LDFLAGS` metadata and less like the runtime environment
capabilities in that there are explicit Plan variables
(`$pkg_bin_dirs[@]`, `$pkg_lib_dirs[@]`, etc.) that map directly to
corresponding metadata files (`PATH`, `LDFLAGS`, etc.) which are used by
the build system to assemble environment variables to be used when
building a package.

This change changes how that `PATH` entry is computed--separately from
the values `RUNTIME_ENVIRONMENT` metadata files. Instead this value is
computed on its own (via `_set_runtime_env_path()`) after all other
environment variable loading is complete and set in the runtime
environment hash which will ultimately be written to the package's
`RUNTIME_ENVIRONMENT` metadata file.

As a result, most of the PATH-related code associated with reading these
environment metadata files was removed, and any `PATH` key entries are
skipped when loading in other package dependencies.

3. Build time value of `$PATH`
------------------------------

In a similar line of reasoning the case 2 above, the build time value of
`$PATH` which gets exported by the build program is calculated
independently of any `RUNTIME_ENVIRONMENT` metadata file values and is
created in the following specific order:

1. All directories in this package which have program binaries (set by
the Plan author in `${pkg_bin_dirs[@]}`), in declared order
2. All path elements for each *direct run* dependency which has a `PATH`
metadata file, in the order present in `PATH`
3. All path elements for each *direct build* dependency which has a
`PATH` metadata file, in the order present in `PATH`
4. All remaining path elements for each *indirect run* dependency which
has a `PATH` metadata file, in the order present in `PATH`
5. All remaining path elements for each *indirect build* dependency
which has a `PATH` metadata file, in the order present in `PATH`
6. The initial/original value of `$PATH` when the program started to
ensure that build program-related programs can still be found

As above, this collection should be a unique ordered set with no
duplicated path elements. The only exception is that we preserve the
initial/original value of `$PATH` to append it as-is (i.e. we don't
de-duplicate the original `$PATH`).

Finally, this task is broken out into a dedicated private phase in the
build (`_set_build_path()`) which happens directly after
`do_setup_environment_wrapper()` and before the `do_before()` build
phase to ensure that `do_before()` implementations have a correctly set
`$PATH`.

---

(use this section for why we introduced a new metadata file)

In the case of PATH, we introduced this really nice-to-have pre-computed
runtime `$PATH` value which the Supervisor and other tooling can use
without have to re-compute this statically-knowable answer at runtime.
This leaves us with 2 notions of PATH: one is used by the build system
and some tooling to answer the question: "where are the programs for
this package?" and the other is used by runtime tools to answer the
question: "what should I prepend to my $PATH before running a program
out of this package?". Notice that these 2 notions are independent of
any other runtime environment variables that a Plan author could
introduce such as `JAVA_HOME`, `PYTHONPATH`, `CA_CERT_FILE`, etc.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>