Fix handling of environment-level configuration
When we were trying to parse TOML from an environment variable, we
weren't properly type-hinting the result of that operation, which
meant Rust derived the type from the return value of our function. As
a result, we ended up with an error, because Serde was trying to
decode directly to an `Option<toml::value::Table>`. We now type things
correctly.

(By the way, the code was all completely valid Rust before; it's just
that Serde is such a flexible framework that can parse into arbitrary
types!)

Additionally, parsing JSON as a fall-back wasn't working because we
were expecting `serde_json` to parse a string into a TOML table
directly, which it can't do. Now, we transcode the JSON string into a
TOML string, and from there use the `toml` crate to parse into a
`toml::value::Table`.

Fixes #4439

Signed-off-by: Christopher Maier <cmaier@chef.io>