This may not be for you; it is for applications
This is almost definitely an irrational pet peeve of mine that is not
rooted in any user research but I really dislike speaking to "you" in
our copy.  This reader might be a manager who does not actually run any
applications, for example.   Or an operations engineer who does not
build applications but does need to run them.

Additionally, I think this reads a bit better.  I do recognize that this
is a pet peeve of mine that I should maybe just let go.

Opening this PR as a way to discuss...

Signed-off-by: Nathen Harvey <nharvey@chef.io>