Add LD_LIBRARY_PATH hack for Supervisor package
This is not needed to build the supervisor, strictly speaking, but is
instead a work-around for how we are currently building packages in
Travis; we hypothesize that the `build.rs` program for
`habitat_http_client`, built during a static hab package build, is
being inadvertently used here. Without gcc libs on the
`LD_LIBRARY_PATH`, the program can't find `libgcc_s.so.1`. This is
merely a bandaid until we can overhaul our CI pipeline properly.

Signed-off-by: Christopher Maier <cmaier@chef.io>