[plan-build-ps1] Treat PATH distinctly from RUNTIME_ENVIRONMENT.
This change updates the PowerShell implementation of the build system to
re-assert how the PATH environment variable is used and interpreted in
different places. Additionally, this change brings back the `LIB_DIRS`
and `INCLUDE_DIRS` metadata files which a package will contain if it has
`$pkg_lib_dirs`/`$pkg_include_dirs` values. Lastly, there is fallback
logic to look in the `RUNTIME_ENVIRONMENT` for `LIB`/`INCLUDE` entries
if the corresponding `LIB_DIRS`/`INCLUDE_DIRS` metadata files aren't
present (this is to handle older built packages which are used as
dependencies in a build).

See the previous change to the Bash implementation for more details and
rationale. The changes here are designed to bring both implementations
in line with each other and continue to produce similar results.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>