Properly honor the `--force` flag
The Boolean here was improperly negated; we want to set it to true if
the user specified `--force` on the CLI, not if they didn't.

Also tweaks some tests to no longer include a package identifier in
the `hab sup run` command, since we've already laid down a spec file
to run. The alternative would be to include `--force`, now that it
works properly :)

Also added a `control_port` to the test supervisor startup command,
just to prevent potential clashes.

Signed-off-by: Christopher Maier <cmaier@chef.io>