Update to current master of habitat-sh/core repository
Some changes were merged in to the core repository
recently (https://github.com/habitat-sh/core/pull/25), with the
anticipation that corresponding changes in this repository would be
merged shortly thereafter. However, those changes are still
outstanding (See #4962).

In the meantime, any additional changes here that require changes to
the `habitat-sh/core` repository will not work until this repository
is updated to deal with the initial changes.

This change updates `Cargo.lock` to point to the current master branch
of `habitat-sh/core`, and makes the necessary compatibility changes to
the `hab` and Supervisor code.

Signed-off-by: Christopher Maier <cmaier@chef.io>