[hab] Ignore subdir entries when binlinking all programs in package.
This change resolves an issue where a user is binlinking all programs in
a package via a `hab pkg binlink` call or by using the `--binlink` flag
in `hab pkg install` and the package has extra subdirectories under
`PATH` entry directories.

For example, let's take a package with a layout such as:

```
.
├── bin
│   ├── dothings
│   └── __pycache__
│       └── dothings.pyc
├── IDENT
├── PATH
└── TARGET
```

If this package only has the `bin/` directory in its `PATH` metadata,
this means that the `bin/__pycache__` directory should be ignored and
skipped over. The implementation prior to this change didn't take this
case into account.

Closes #4847
Closes #4669

References #4390

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>