Fix Rust 1.25 compilation errors
In #5090 and #5091, validation of configuration application was
temporarily disabled, pending some more complete changes.

In the process, a compilation error was introduced on Rust 1.25. Some
developers are using 1.26, but TravisCI is currently pinned to
1.25. The code in question relied on implicit type inference rules
that are slated to change. The fix here was simple (just type-hint the
result), and can be removed once we update Travis' Rust compiler
toolchain.

Unnecessary mutability and unused variable warnings were fixed, as
well.

Signed-off-by: Christopher Maier <cmaier@chef.io>