Fix default configuration serialization bug
TOML serialization is picky, in that table keys that are themselves
keys must be rendered last in the table.

We have typed our TOML table data as `toml::value::Table`, which is
just an alias for a `BTreeMap`. However, the smarts for serializing
tables are attached to the `toml::value::Value` enum, where
`toml::value::Table` is the payload of the `toml::value::Value::Table`
variant.

I'm not sure that we want to retype everything, though, since you
can't constrain things to a single Enum variant (everything that is
definitely a `Table` right now would end up just being a
`Value`). Longer term, we may want to look at creating our own
concrete types, but for now, we can wrap our default configuration
with `toml::value::Value::Table` to get the correct serialization
logic.

Fixes #5109

Signed-off-by: Christopher Maier <cmaier@chef.io>